Create buffered ride polygons
In order to improve the on hover behavior for a ride polyline, we need
to buffer it with a polygon. We can then assign component state when
the invisible buffered polygon is hovered over, and use this
component state to highlight the polyline and render the ride
information below the map.